fixed the actual ruin rate calculation (integer divison was rounding to 0; changed values to doubles which corrected the issue)

+++ src/studio2/Ruin.java
@@ -49,7 +49,7 @@ public class Ruin {
             expectedRuinRate = (Math.pow(a, startAmount) - Math.pow(a, winLimit))/1-Math.pow(a, winLimit);
         }
 
-        ruinRate = losses/totalRuns;
+        ruinRate = (double)losses/totalRuns;
 
         System.out.println("Losses: " + losses);
         System.out.println("Simulations: " + totalRuns);
